+++ test/renderUtilsForTests.ts
@@ -1,0 +1,8 @@
+import { RenderUtils } from '../src/renderer/renderUtils';
+export default function(): RenderUtils {
+  return {
+    publicUrl(_imageUrl: string): string | null {
+      return 'https://dev.null';
+    },
+  };
+}