+++ CHANGELOG.md
@@ -3,7 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## 1.0.0
+## 1.0.0 - 2020-01-14
+
+### Fixed
+- restore all the unit tests
+- handle (by skipping) the toggle block type
+
+## 1.0.0 - 2020-01-12
 ### Added
 - add an example in `examples` directory
 - publish site at http://www.gatsbyplugins.com